fix: handle timeout on SFU WS connections (#1600)
### Fixes hanging join requests when SFU WS connection cannot be
established

Previously, `signalReady` for SFU WS never rejected if WS connection
couldn't be established. We now reject it after a timeout, so that the
`call.join()` doesn't hang forever.

We reuse the `joinTimeout` option as a sensible default here.

### Adds support for disconnection timeout setting

When going through the reconnection flow, we now take into account the
new `disconnectionTimeoutSeconds` setting, so we don't retry
reconnecting indefinitely.