fix: use worker to prevent timer throttling (#1557)
Browsers start throttling timers (intervals, timeouts) in background
tabs after a period of inactivity. We use timers to send health checks,
and missing health checks cause participant to be disconnected.

An active WebRTC connection on a page usually means the page is never
considered inactive by the browser, and therefore timers are not
throttled. But it's still possible to be in a call and _not_ have an
active connection, e.g. if nobody is publishing neither audio nor video.
Also, some browsers (like Edge) might be more aggressive about timer
throttling.

To handle situations like this, we move timers to a web worker, which
are never throttled. Web worker then posts messages to the page on the
specified intervals.

This feature is currently opt-in and hidden behind a feature toggle,
which can be enabled when creating a client instance. Also, web workers
might not be available (banned by CSP, running on Node, etc), so there's
also a fallback mode that uses regular timers on the page side even when
enabled.