fix: adjust dynascale debouncing for upscaling and downscaling (#1621) [skip ci]
We currently debounce requests to update participant subscriptions when
a bound video element size changes with a delay of 1200ms. That's too
long of delay, but there are also other considerations.

Not all size changes are equal: when upscaling, low quality video in a
larger video element is very noticeable; high quality video in a smaller
video element for a second or two is not a big deal.

Proposal:
1. When a video element upscales significantly (currently that means
that either width or height went up more than 20%), update participant
subscription with minimal debouncing.
2. When a video element upscales a little, debounce more aggressively,
to guard against quick-firing resize events.
3. When a video element downscales, it's also fine to debounce more
aggressively. That will guard us against switching to lower quality when
several people are talking over other and the dominant speaker changes
quickly.