feat: [VID-315] allow filtering participants using filter object (#1655)
This PR adds support for filtering participants using a filter object
instead of a predicate. Filter object looks something like this:

```js
// Render video only from pinned participants
// or participants with the role "presenter"
{
  $or: [
    { isPinned: true },
    { roles: { $contains: 'presenter' } },
  ],
}
```

(For more examples of what kinds of operators we allow, see
`filters.test.ts`.)

Filtering is supported for a subset of participant properties:
- userId: string
- isSpeaking: boolean
- isDominantSpeaker: boolean
- name: string
- roles: string[]
- isPinned: boolean

`SpeakerLayout` and `PaginatedGridLayout` now accept filter objects in
their `filterParticipants` prop:

```jsx
<SpeakerLayout
  filterParticipants={{
    $or: [
      { roles: { $contains: 'student' } },
      { isPinned: true },
    ],
  }}
/>
```

This is particularly useful for the composite app, because a filter
object (unlike a predicate) can be passed in configuration. So the new
optional configuration option is added to the composite app:
`participant.filter`. It is supported by built-in layouts.