fix: ensure tracks are stopped when disposing a Publisher (#1676)
In some circumstances, we were dangling track clones. No data flowed to
the SFU, but the browser kept the camera/mic indicators on as there were
MediaStreamTrack instances that weren't stopped properly.

### Implementation notes

Upon disposing of a `Publisher`, we now iterate through all transceivers
and we explicitly stop the assigned track.
This PR also fixes another issue that could lead to dangling, unstopped
tracks when replacing a track on an existing transceiver.