fix: voip token not able to be got if user switched (#1683)
### Overview

Consider this scenario where as soon as app is opened. User A logged in
and then immediately logged off and then user B logs in.

```
1. user A logs in
2. useEffect registers listeners to listen for push
3. user A logs out immediately - listeners are removed
4. apple sends push token now , but SDK is not listening as no user is logged in
5. user B logs in
6. useEffect registers listeners to listen for push
7. apple DOES NOT SEND push token now because it was sent earlier
```

### Implementation notes

If we look at the voip notification library, if we call
registerVoipToken() method, the library sends the JS event for new push
token with the last push token that it received.

Now after this in point 7 we will receive a event in JS saying that it
received a push token