feat: support static token and token provider at the same time (#1685)
Until now, providing both, `token` and `tokenProvider` wasn't truly supported, and when this was the case, the `client` always used the`static` token.

This PR allows the client to connect faster and save one roundtrip to the customer's servers.

The logic is the following:
- if only `token` is provided - use the `token` (existing behavior)
- if only `tokenProvider` is provided - use the `tokenProvider`
(existing behavior)
- if both `token` and `tokenProvider` are provided - use the `token`
initially. Once it expires, use the `tokenProvider` (new behavior).

The new behavior aligns with the logic that the other SDKs have, and it
potentially reduces the time it takes to establish a WS connection to
our systems.

In addition, we now validate internally that the token provided by the
token provider is issued for the correct `user_id.` This should prevent
potential integration issues.

This PR also refactors and cleans the code used to construct a new `StreamVideoClient` and also adds test cases for the critical parts of it.

Also, a few dependencies are updated to their latest versions (`axios`,`sdp-transform`, `ua-parser-js`, `vitest`, etc...).