fix: don't implicitly mark calls as `ringing` (#1697)
Fixes an issue where reused ringing `call` instances were implicitly marked as `ringing`.

For example:
```ts
await call.get({ ring: true });
await call.leave();

// This is used to mark the `call` instance as `ringing` 
// and set the `callingState` to `RINGING`
await call.get();
```

### Implementation notes
We don't consider the (past) internal state of the `call` instance when
it is reinitialized. This aligns the behavior with the rest of the SDKs.
- https://getstream.slack.com/archives/C07EV8DFMRN/p1740394903388349

Fixes:
https://github.com/GetStream/stream-video-js/issues/1561#issuecomment-2662584543