fix: do not stop incall manager on change of audio settings (#1702)
If call settings are changed on the fly when a call is already joined,
then incall manager is started and stopped again.

InCallmanger mode with "audio" is almost never used by our integrators.
Speaker is the default preferred output in any scenario. We should not
route to earpiece in video disabled calls by default too. I do not know
why this was added.
 
This behaviour also aligns with other SDKs and also fixes the bug on iOS
that audio is lost after start->stop->start