fix: add explicit default device option to device selectors (#1701)
As far as we know, there's no reliable way to determine the system's
default camera and microphone devices without querying them first-and we
shouldn't query the user's media devices unless they're enabled. So we
have no way of knowing which camera or microphone device is being used
by the browser under the following conditions

1. No device has been explicitly selected by the user (or the previously
selected device is no longer available).
2. No device was previously enabled by the user

Previously, in this case, we simply rendered a list of devices with no
device selected. This was confusing and looked "broken".

Now we add an explicit "Default" option to the list instead. This option
is purely cosmetic: if it's present, it's always selected, so selecting
it does nothing; and if a different device is selected, this option
disappears.

![Screenshot 2025-02-26 at 16 26
02](https://github.com/user-attachments/assets/dae2d549-4f5e-4616-8b34-e01acd26e0de)

![Screenshot 2025-02-26 at 16 25
29](https://github.com/user-attachments/assets/3d47a33a-cd79-4fa6-9279-af925e0281c9)

### Implementation notes

This is implemented on the React SDK level, and the new `useDeviceList`
hook is applied to device lists in all device selectors.

This hook is public API, documented here:
https://github.com/GetStream/docs-content/pull/206