fix: retryable client.connectUser() (#1710)
Adds a retry mechanism for the `client.connectUser()` API.
By default, the SDK will now attempt to establish the initial WebSocket connection up to 5 times.
Once all attempts are exhausted, and are unsuccessful, the `connectUser()` will reject with the last captured error as a rejection reason.

### Implementation notes
Connection flow can happen in the `StreamVideoClient` constructor, but
due to its non-async nature, integrators can't catch and handle these
errors. This is a flaw that we introduced early in the development of
the SDK and fixing it is a huge breaking change. We need to consider
fixing this when we start planning for the v2 release of the SDK.
Until then, a new `onUserConnectError` callback can be provided in the
constructor's options that will be invoked once the SDK gives up from
retrying. This callback will get the `lastError` and the `allErrors`.

```ts
const client = new StreamVideoClient(apiKey, {
  defaultWsTimeout: 5000,
  maxConnectUserRetries: 3,
  onConnectUserError: (lastError: Error, allErrors: Error[]) => {
    // handle connect errors
  },
});

// alternatively
const client = StreamVideoClient.getOrCreateInstance({
  apiKey,
  user,
  options: {
    defaultWsTimeout: 5000,
    maxConnectUserRetries: 3,
    onConnectUserError: (lastError: Error, allErrors: Error[]) => {
      // handle connect errors
    },
  },
});
```