fix: ensure negotiation runs sequentially (#1722)
When negotiating, we shouldn't run two SetPublisher requests in parallel
as they can arrive in non-deterministic order on our server.
Sometimes it used to happen that we ran two SetPublisher requests:
1. Audio
2. Audio + Video

When 2. arrives first, then 1 will overwrite the announced video track
and will lead to issues where participant A can't see the video of
participant B.

This PR aims to fix that by ensuring that we wait for Req-1 to finish
before we run Req-2.

This PR also fixes another issue, where `codecInUse` could have been
incorrectly determined when handling the `changePublishQuality`.

Ticket:
https://linear.app/stream/issue/REACT-310/ensure-sequential-negotiation

### Implementation notes
- We guard `negotiate()` with a concurrency lock that's shared between
both, ICE restarts, and regular negotiations
- We don't use `negotiationneeded` event anymore - we explicitly control
the negotiation process that enables better exception handling, as we
can now bubble up the negotiation errors to the original initiator