fix: proper non ringing only push support without ringing libs (#1731)
### Overview

Currently when only non ringing push is configured and ringing is not.
We still check for ringing push deps. This is fixed in this PR.

Also, on expo when only ringing was configured we still checked for
expo-notifications presence. This should not be the case. Also fixed in
this PR.


Extra: some logs added for debugging purposes