fix: screenshare_layout is not injected in egress app (#1754)
### 💡 Overview

The `screenshare_layout` option (that supposedly controls which layout
is used for recording calls with ongoing screen shares) is not injected
into egress app.

We now fallback to `layout` when `screenshare_layout` is missing,
instead of Spotlight layout.

### 📝 Implementation notes

In general, seems like `screenshare_layout` option is not needed,
because all layouts implement default view and screenshare view.