fix: fast reconnect shouldn't be followed up with full rejoining on network switch (#1760)
We weren't properly handling the Fast reconnects on the network switch.
Every fast reconnect was later followed up with a full rejoin due to the
`previousSfuClient` is not being correctly flagged as disposed of.
This behavior wasn't an issue that had a destructive impact on end users
but it did create multiple timelines in our dashboard stats (due to the
`session_id` change) and made it more difficult for us to investigate
call issues and assign call scores.

🎫 Ticket: https://linear.app/stream/issue/REACT-354