fix: access device list lazily from call state hook (#1761)
### 💡 Overview

Previously, rendering call state hooks like `useCameraState` and
`useMicrophoneState` triggered permission prompt. That happened because
these hooks access device list.

Some customers rightly find this confusing: rendering `useCameraState`
doesn't necessarily mean you are going to access device list.

Now, device list is accessed only when the `devices` property returned
by the hook is really read. If it is not read, no permission prompt will
be triggered by the hook.

### 📝 Implementation notes

The `devices` property is now a getter that forces component update when
accessed for the first time. Most of the times it will happen during
render, which is OK.

> **Note**
> 
> There's one edge case when updating state from getter will not work:
that is, if you save hook value to a variable (`const state =
useCameraState()`) and then pass it to another component
(`<SomeComponent state={state} />`), and then this component accessed
device list (`function SomeComponent(props) { props.state.devices }`).
> 
> It's safe to assume nobody will every do that.

Of course this means double-rendering component when the device list is
first accessed. But at least it happens synchronously, without a paint
in between.