fix: call cancellation not reliably seen on killed state RN-198 (#1773)
### 💡 Overview

Sometimes this happens

```
 when A initiates a call to B and then A cancels the call but the notification has already reached B and the notification is not cleared at B’s side.
```

Technically, this happens when StreamCall is mounted after Foreground
service for incoming call. This overwrites the foreground service
responsible for listening to call cancellation.

### 📝 Implementation notes

In this PR, we delay the foreground service setting for keep call alive
only when call is joined so that we dont overwrite the foreground
service.

🎫 Ticket:
https://linear.app/stream/issue/RN-198/notification-not-cleared-on-call-cancellation-sometimes-android

📑 Docs: NA