fix: make camera.flip() work more reliably with older devices (#1781)
While optimizing the `camera.flip()` API in #1679, we introduced a
regression that led to `camera.flip()` throwing a `NotReadableError` on
devices with limited hardware capabilities.
Some devices are unable to provide two `MediaStream` instances at the
same time.

### 📝 Implementation notes
We want to preserve the `camera.flip()` optimizations as they allow
significantly faster direction change.
We always attempt to perform a fast direction change, but if it fails,
we fall back to:
`stop()` -> `flip()` -> `start()` flow, which is slower, but it works.

🎫 Ticket: https://linear.app/stream/issue/REACT-367/