fix: isolate mediaDevices traces (#1779)
Continuation of #1765. `mediaDevice` traces were tracked in a single
shared `Tracer` instance.
While this is ok for the majority of the integrations, this can be
problematic for customers allowing one to be connected to two separate
`call` instances from a single browser tab.
This PR scopes the traces to the `call` instance.

### 📝 Implementation notes
- We aren't patching browser built-in mediaDevice APIs anymore
- We are drilling down a `tracer` instance in the lowest level APIs
- We can't capture `getUserMedia` invocations triggered outside of our
SDK anymore
- Updates the format of the `create` trace - we include the SFU name in
it now

🎫 Ticket:
https://linear.app/stream/issue/REACT-365/rtc-tracing-and-performance-stats-tracking