fix: use scoped locking for PeerConnection events (#1785)
We have locks to ensure sequential processing of certain Peer
Connection-related events.

While this works well in the happy path, it sometimes fails when there
are multiple `call` instances running in parallel.
e.g., processing an event for Call A halts the processing of the events
meant for Call B.
The real problem happens when Call A halts completely and causes a
complete stop of processing events for any other Call instance. This is
because we keep a global Promise chain that isn't scoped.

This bug affects only the few `withoutConcurency(string)` usages we have
throughout the SDK; .symbols are unaffected.

### 📝 Implementation notes
- Add a random chain identifier for the PeerConnection implementation
- We don't use `call.cid` as there could be multiple `call` instances
for the same `cid` too

🎫 Ticket: https://linear.app/stream/issue/REACT-375