fix: fixes an edge case where tracks weren't restored after a reconnect (#1789)
Sometimes when the WS was abruptly closed, the local track state wasn't
restored properly until the user toggled the mic/camera.
The reason for this was improper handling of sending the `LeaveRequest`
to the broken SFU.
The `doJoin()` method never returned, as it was awaiting the "broken
client" to reconnect, which obviously won't happen.
This led to a situation where `restorePublishedTracks()` and
`restorePublishedTracks()` never ran, and kept the UI state
inconsistent.

Adds support for `useSingleLayer` publish option that which enabled,
would start simulcast with only one layer enabled.
This should reduce the chances of layer throttling at join time.

### 📝 Implementation notes
- Skip sending the `LeaveRequest` if the previous WS connection is not
healthy
- Add handling for `useSingleLayer`
- Flush stats on `call.leave()`
- Improved handling of `failed` ICE state -> immediately trigger a
rejoin once this happens