fix: do not mutate filters array during pipeline setup (#1798)
It's a possible fix for the issue when there's no audio coming out from the noise cancellation audio context.

### 📝 Implementation notes

Our filter pipeline setup is async, so we should be careful not to
mutate anything filter-related while the pipeline is being set up. Even
if it's not a root cause of the problem described above, it can still
lead to weird consequences.