fix: rpc error tracing (#1801)
When an error happens while processing an RPC, our SFU returns an HTTP
200 status and a response body with an `error` object. This is a bit of
non-standard behavior that we can't fix without breaking the existing API.
This means that data errors will never throw an error, and hence, making
our try/catch interceptor useless in some circumstances.
In this PR, we unwrap the response in the interceptor (asynchronously)
and inspect its content.