feat(android): Accept incoming call without device unlock (#1806)
### 💡 Overview

*Main goal*: adding ability that phone need not be unlocked all the time
when call is joined

*History*: currently, when incoming call arrives and if full screen
intent permission is added to app. The incoming call screen is shown,
but then after phone needs to be unlocked.

### Implementation

Uses
[requestDismissKeyguard](https://developer.android.com/reference/android/app/KeyguardManager#requestDismissKeyguard(android.app.Activity,%20android.app.KeyguardManager.KeyguardDismissCallback))
method from Android platform. Here is the behaviour as per the doc

```
If the Keyguard is not secure or the device is currently in a trusted state, calling this method will immediately dismiss the Keyguard without any user interaction.

If the Keyguard is secure and the device is not in a trusted state, this will bring up the UI so the user can enter their credentials. 
```

📑 Docs: https://github.com/GetStream/docs-content/pull/358