fix: use AudioContext for Safari (#1810)
### 💡 Overview

Safari's power-management and autoplay policy do not allow a tab to
start playing audio through an `<audio />` element when the user doesn't
have the tab in focus. This behavior causes an issue where a remote
participant can't be heard until the local participant focuses on the
tab that has the ongoing call.

Steps to reproduce:
- The local participant switches to a different tab
- The remote participant mutes and unmutes while the local participant
is browsing on a different tab
- The remote participant becomes audible only after the local
participant switches back to the call tab

### 📝 Implementation notes
- Uses `AudioContext` for Safari, as it doesn't seem to be restricted by
the above-mentioned policies
- The rendered `audio` elements are now muted by default on Safari, as
we want to prevent "double audio"
- The implementation is scoped to Safari only, but we should consider
moving fully to it in a future major release, and reduce the amount of
`<audio>` elements our SDK renders

🎫 Ticket: https://linear.app/stream/issue/REACT-405

---------

Co-authored-by: Matvei Andrienko <m.y.andrienko@outlook.com>