fix: improved video quality on low capture resolution (#1814)
Improves the video quality on low capture resolution.
When capturing with relatively low resolution (640x480), we were too aggressively downscaling the `q` layer to 160x120.
Although this looks good in a grid layout with small tiles, it creates quite a pixelated video on 1:1 calls.
In this PR, we reduce the aggressiveness and downscale to 320x240 - a resolution that looks sharper on smaller tiles and improves the pixelation.

This PR also fixes an issue with our implicit requirement that `target_resolution` must be in _landscape_ format.
Adds more telemetry data tracing, too.

🎫 Ticket: https://linear.app/stream/issue/RN-224