feat: add audio only ringing notification support for iOS  (#1821)
### 💡 Overview

Callkit gives a way to indicate audio only calls. This is supported by
our backend.

In this PR, we add it to expo plugin and dogfood app


![IMG_0006](https://github.com/user-attachments/assets/69c5d30e-4aa4-486d-b852-6125b310b9b2)

Notice the audio ⬆️ 

### 📝 Implementation notes

This pull request introduces changes to handle the presence of video in
incoming calls more dynamically by checking the `video` field in the
incoming stream data. The updates ensure that the `hasVideo` parameter
is set based on the value of the `video` field instead of being
hardcoded.

* **Swift updates in `withAppDelegate.ts` and `AppDelegate.swift`:**
- Added logic to parse the `video` field from the incoming stream data
and determine the `hasVideo` value dynamically.
- Updated `RNCallKeep.reportNewIncomingCall` to use the dynamically
determined `hasVideo` value instead of a hardcoded `true`.

* **Objective-C updates in `withAppDelegate.ts`:**
- Introduced logic to evaluate the `video` field and set the `hasVideo`
value accordingly.
- Modified `RNCallKeep.reportNewIncomingCall` to utilize the dynamically
calculated `hasVideo` value. (

📑 Docs:
[https://github.com/GetStream/docs-content/pull/<id>](https://github.com/GetStream/docs-content/pull/375)