fix: correctly setup and dispose device managers (#1836)
### 💡 Overview

We had a discrepancy in how device managers were created and disposed.

Device manager is created in Call constructor, and is available for the
whole lifetime of the call instance.

When created, device manager also sets up some subscriptions. Some of
these subscriptions only really work within a fully initialized call
(after `call.get()`, `call.join()` etc). Some, like
`handleDisconnectedOrReplacedDevices` are nice to have even if the call
is not initialized.

When leaving call, all subscriptions are removed in the device manager's
`dispose()` method. However, the call can be joined again after leaving,
and device manager subscriptions were not set up again in this case.

This PR fixes that by setting up device manager subscriptions again when
the call is initialized again after leaving.

**Sidenote:** device config was also not applied again when joining the
call after leaving because the `deviceSettingsAppliedOnce` flag was not
reset on leaving the call. This is also fixed in this PR.

### 📝 Implementation notes

The solution in this PR is not ideal. The reason is that we don't really
have a method to dispose call instance. Instead, `leave()` _kind of_
stops most running side effects, but the call instance can be then
revived by calling `get()`, `join()`, `create()` etc.

To keep supporting this behavior, we set up device manager subscriptions
twice:
1. When device manager is created
2. When call is initialized within the `call.setup()` method

That allows us to stop side effects in device manager when call is left,
but also to run them again if the call instance is reused.

There's a flag `areSubscriptionsSetUp` in place to prevent duplication
of subscriptions.

🎫 Ticket:
https://linear.app/stream/issue/REACT-428/camera-and-mic-not-enabling-when-rejoining-call-in-web-client