fix: resolve `default` device id into real id (#1839)
### 💡 Overview

As we are well-aware at this point, Chromium-based browsers report a
special device with id `default` for audio input and output. For
cross-browser consistency, we filter out this device from all our device
lists.

However, Chrome can also have `default` as the device id in
MediaStream's audio track settings. When this happens, we should replace
`default` with the real device id.

### 📝 Implementation notes

We rely on `groupId` to find the "real" device that corresponds to the
`default` one. If we find a device with the same `kind`, same `groupId`
as the `default` one, and if it's the only device like this, we can
safely replace `default` with it.