fix: sync call state after a failed reconnect (#1846)
When a call reconnect fails because of `unrecoverable` error (typically
due to a permission revocation, or call being ended in the meantime),
the client-side call state wasn't up-to-date with the server-side one
(`/join` didn't return any data).

### 📝 Implementation notes
In this PR, before flagging the `callingState` as `RECONNECTING_FAILED`,
we perform an additional request to our backend so we can synchronize both states.

🎫 Ticket:
https://linear.app/stream/issue/RN-241/refresh-call-state-after-a-failed-reconnect