feat: pronto for external calls (#1873)
### 💡 Overview

We are adding a separate environment to pronto, used for external calls.
In this environment, users have less "dangerous" permissions, like
querying calls or recording. Stream employees can login in via Google
with their getstream.io emails.

The app also has some integrations set up: creating a meeting from
Google Calendar; uploading call recordings to Gong.

### 📝 Implementation notes

**Authentication.** When authenticating with Google, the session
contains `stream: true` flag. When creating a Stream token for the
session, the user with role `stream` is also created.

Sign in page has a new `signIn=1` optional flag to immediately sign in
as a non-Stream user. This allows to create a token for immediately,
without additional steps.

**Display name.** We now allow changing user name in lobby. This doesn't
affect the session, it just changes Stream user name.

**Google Calendar integration.** The new integration supports syncing
meeting name with `custom.name` property on the call. The calendar
add-on will call the `/gcal/call-id` endpoint to update the call.

**Gong integration.** Call recording are uploaded for processing to
Gong. When creating a conversation in Gong, we use the first call member
with `stream` role as the primary user. (Every user that joins the call
becomes a member, see the `saveParticipantAsCallMember` webhook.)

If the integration flakes out, refer to webhook logs on Vercel to see
what went wrong.