fix: support setting iOS pip window sizes (#1876)
### 💡 Overview

iOS pip window did not know what size to set. And so size was set by the
OS. This was not ideal as it seemed to completely random at times. In
this PR we add support for PiP window size based on track size.



https://github.com/user-attachments/assets/0875304c-b812-4fc6-b74b-78d3c48c4fd1


### 📝 Implementation notes

Sends track dimensions from JS to Native