fix: respect call type settings when applying persisted device preferеnces (#1879)
Before applying local preferences, we now check whether the appropriate
device kind is enabled on the call type level.
This is especially important for customers using default and audio-only
call types in the same app.

Although preferences can be kept under a different key depending on the
call type, it is good to handle this in the hook itself, too.