fix: synchronize ring events (#1888)
The order of `call.created` and `call.ringing` can't be guaranteed.
Until now, the SDK assumed that `call.created` would arrive first, but
we've seen that's not always the case.
When the order was flipped, the SDK had a race condition that made the
`ringing` call be created in `non-ringing` mode, causing unexpected
integration issues.

### 📝 Implementation notes

This PR unifies the handling of both events and correctly synchronizes
the access to them.
Also, adds a few tests so future regressions are prevented.

This PR also fixes a bug in our Pronto ringing integration. The caller
should always be included in the `members` array.

TODO: There is one failing test, which I believe is a backend bug. I'm
waiting on our BE team's feedback.

🎫 Ticket:
https://linear.app/stream/issue/REACT-528/synchronize-callring-events