fix: don't apply default camera state if video is off (#1917)
For the calls with video disabled (e.g. when video is disabled for the
call type), the `camera_default_on` setting should be ignored.
Currently, with these settings the SDK tries enabled camera when the
call is set up:

```json
{
  "video": {
    "enabled": false,
    "camera_default_on": true
  }
}
```

This results in an error log in browser console and camera's optimistic
state switching unnecessarily.