feat(react-native): reject call when busy (#1856)
### 💡 Overview

Adds support in the react-native SDK for automatically rejecting a call
with reason `busy` when the callee is in another call, so that
integrators can implement handing of the `callee busy scenario`.

### 📝 Implementation notes

Adds `shouldRejectCallWhenBusy` props in React Native push notifications
configuration. (`false` by default)
Adds ios and android native methods for play and pause busy tone. 

Handles the call rejection case in the sample app with:
  - showing toast message: `Call rejected because user is busy.` 
  - playing busy tone using native methods


🎫 Ticket: https://linear.app/stream/issue/RN-232

📑 Docs: https://github.com/GetStream/docs-content/pull/476

---------

Co-authored-by: Santhosh Vaiyapuri <3846977+santhoshvai@users.noreply.github.com>
Co-authored-by: Santhosh Vaiyapuri <santhoshvai@gmail.com>